pyproject.toml
[project]
name = "vira"
version = "0.1.0"
description = "Venture Intelligence Research Assistant - Iteration 1 prototype"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "langchain>=0.2.14",
  "langchain-community>=0.2.11",
  "langchain-openai>=0.1.21",
  "chromadb>=0.5.3",
  "rank-bm25>=0.2.2",
  "scrapy>=2.11.1",
  "beautifulsoup4>=4.12.3",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.3.4",
  "PyMuPDF>=1.24.9",
  "python-docx>=1.1.2",
  "langchainhub>=0.1.21,<0.2.0",
  "tiktoken>=0.7.0",
  "streamlit>=1.36.0",
  "httpx>=0.27.0",
  "tenacity>=8.5.0",
  "orjson>=3.10.7",
  "typer>=0.12.5",
  "jinja2>=3.1.4",
  "chainlit>=1.0.0",
  "sqlalchemy>=2.0.0",
  "aiosqlite>=0.19.0",
  "websockets>=12.0",
  "python-multipart>=0.0.6",
  "langgraph>=0.2.0",
  "google-search-results>=2.4.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "mypy>=1.11.1",
  "ruff>=0.6.4",
  "black>=24.8.0",
  "ipykernel>=6.29.5"
]

[project.scripts]
vira-crawl = "vira.ingestion.runner:main"
vira-process = "vira.processing.cli:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B", "UP", "N", "ANN", "C4"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
ignore_missing_imports = true

